Move encryption steps to EncryptionContext to be like the encryption app
