+++ tests/acceptance/config/behat.yml
@@ -47,6 +47,7 @@ default:
         - '%paths.base%/../features/cliEncryption'
       contexts:
         - FeatureContext: *common_feature_context_params
+        - EncryptionContext:
         - OccContext:
         - WebDavPropertiesContext:
 
@@ -259,6 +260,7 @@ default:
         - WebUIGeneralContext:
         - WebUILoginContext:
         - FeatureContext: *common_feature_context_params
+        - EncryptionContext:
         - OccContext:
         - WebDavPropertiesContext:
